embeddings: use NewEmbedder for Vertex embeddings
For #379
